requirements.txt
nltk
numpy
tensorflow
Pillow
